refactor: apply empirically validated comprehension improvements
Keep only the code refactors validated in the comprehension autoresearch pass, excluding autoresearch harness/log artifacts.

Highlights:
- consolidate shared CLI parsing, client, resolver, and fetch helpers
- flatten large command handlers and split oversized command files
- split large package files in client, exchange, info, and resolver into focused units
- remove duplicated command-layer validation already enforced downstream
- simplify request/result assembly while preserving behavior

Validation: make check
