+++ cmd/account.go
@@ -3,18 +3,12 @@ package cmd
 import "github.com/spf13/cobra"
 
 func newAccountCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "account",
-		Short: "Account transfers, withdrawals, and agent management",
-		Long: `Transfer USDC between spot and perp, withdraw to Arbitrum, manage agent
+	return newHelpCommandGroup(
+		"account",
+		"Account transfers, withdrawals, and agent management",
+		`Transfer USDC between spot and perp, withdraw to Arbitrum, manage agent
 wallet approvals, and perform cross-account transfers. Account commands
 sign with the configured private key via the user-signed path (chain ID 421614).`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return cmd.Help()
-		},
-	}
-
-	cmd.AddCommand(
 		newAccountTransferCmd(),
 		newAccountWithdrawCmd(),
 		newAccountClassTransferCmd(),
@@ -22,6 +16,4 @@ sign with the configured private key via the user-signed path (chain ID 421614).
 		newAccountApproveAgentCmd(),
 		newAccountSetAbstractionCmd(),
 	)
-
-	return cmd
 }
