+++ cmd/account_set_abstraction.go
@@ -1,21 +1,11 @@
 package cmd
 
 import (
-	"strings"
-
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/spf13/cobra"
 
 	"github.com/timbrinded/hlgo/pkg/config"
 	"github.com/timbrinded/hlgo/pkg/exchange"
-	"github.com/timbrinded/hlgo/pkg/output"
 )
-
-var allowedAbstractions = map[string]struct{}{
-	"unifiedAccount":  {},
-	"portfolioMargin": {},
-	"disabled":        {},
-}
 
 func newAccountSetAbstractionCmd() *cobra.Command {
 	cmd := &cobra.Command{
@@ -26,30 +16,13 @@ func newAccountSetAbstractionCmd() *cobra.Command {
 			user, _ := cmd.Flags().GetString("user")               //nolint:errcheck // known flag
 			abstraction, _ := cmd.Flags().GetString("abstraction") //nolint:errcheck // known flag
 
-			if !common.IsHexAddress(user) {
-				return output.NewCLIError(output.ErrValidation, "invalid user address").
-					WithDetails("user", user)
-			}
-			abstraction = strings.TrimSpace(abstraction)
-			if abstraction == "" {
-				return output.NewCLIError(output.ErrValidation, "abstraction is required")
-			}
-			if _, ok := allowedAbstractions[abstraction]; !ok {
-				return output.NewCLIError(output.ErrValidation, "unsupported abstraction value").
-					WithDetails("value", abstraction).
-					WithDetails("allowed", []string{"unifiedAccount", "portfolioMargin", "disabled"})
-			}
-
 			exec, err := buildExecutor(cfg)
 			if err != nil {
 				return err
 			}
 
-			raw, err := exec.UserSetAbstraction(cmd.Context(), exchange.UserSetAbstractionInput{
-				User:        strings.ToLower(user),
-				Abstraction: abstraction,
-				DryRun:      cfg.DryRun,
-			})
+			input := exchange.UserSetAbstractionInput{User: user, Abstraction: abstraction, DryRun: cfg.DryRun}
+			raw, err := exec.UserSetAbstraction(cmd.Context(), input)
 			if err != nil {
 				return err
 			}
@@ -61,10 +34,7 @@ func newAccountSetAbstractionCmd() *cobra.Command {
 	cmd.Flags().String("user", "", "user address")
 	cmd.Flags().String("abstraction", "", "abstraction string")
 
-	for _, required := range []string{"user", "abstraction"} {
-		//nolint:errcheck // MarkFlagRequired on known flags never fails
-		cmd.MarkFlagRequired(required)
-	}
+	mustMarkRequiredFlags(cmd, "user", "abstraction")
 
 	return cmd
 }
