+++ cmd/account_transfer.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"github.com/shopspring/decimal"
 	"github.com/spf13/cobra"
 
 	"github.com/timbrinded/hlgo/pkg/config"
@@ -10,9 +9,19 @@ import (
 )
 
 func newAccountTransferCmd() *cobra.Command {
+	return newUSDClassTransferCmd(
+		"transfer",
+		"Transfer USDC between spot and perp classes",
+		"USDC amount",
+		"transfer from spot to perp",
+		"transfer from perp to spot",
+	)
+}
+
+func newUSDClassTransferCmd(use, short, amountHelp, toPerpHelp, toSpotHelp string) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "transfer",
-		Short: "Transfer USDC between spot and perp classes",
+		Use:   use,
+		Short: short,
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			cfg := config.FromContext(cmd.Context())
 			amountStr, _ := cmd.Flags().GetString("amount") //nolint:errcheck // known flag
@@ -23,10 +32,9 @@ func newAccountTransferCmd() *cobra.Command {
 				return output.NewCLIError(output.ErrValidation, "exactly one of --to-perp or --to-spot is required")
 			}
 
-			amount, err := decimal.NewFromString(amountStr)
+			amount, err := parseDecimalField("amount", amountStr)
 			if err != nil {
-				return output.NewCLIError(output.ErrValidation, "invalid amount").
-					WithDetails("value", amountStr)
+				return err
 			}
 
 			exec, err := buildExecutor(cfg)
@@ -34,11 +42,8 @@ func newAccountTransferCmd() *cobra.Command {
 				return err
 			}
 
-			raw, err := exec.USDClassTransfer(cmd.Context(), exchange.USDClassTransferInput{
-				Amount: amount,
-				ToPerp: toPerp,
-				DryRun: cfg.DryRun,
-			})
+			input := exchange.USDClassTransferInput{Amount: amount, ToPerp: toPerp, DryRun: cfg.DryRun}
+			raw, err := exec.USDClassTransfer(cmd.Context(), input)
 			if err != nil {
 				return err
 			}
@@ -47,11 +52,10 @@ func newAccountTransferCmd() *cobra.Command {
 		},
 	}
 
-	cmd.Flags().String("amount", "", "USDC amount")
-	cmd.Flags().Bool("to-perp", false, "transfer from spot to perp")
-	cmd.Flags().Bool("to-spot", false, "transfer from perp to spot")
-	//nolint:errcheck // MarkFlagRequired on known flags never fails
-	cmd.MarkFlagRequired("amount")
+	cmd.Flags().String("amount", "", amountHelp)
+	cmd.Flags().Bool("to-perp", false, toPerpHelp)
+	cmd.Flags().Bool("to-spot", false, toSpotHelp)
+	mustMarkRequiredFlags(cmd, "amount")
 
 	return cmd
 }
