+++ cmd/info.go
@@ -3,18 +3,12 @@ package cmd
 import "github.com/spf13/cobra"
 
 func newInfoCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "info",
-		Short: "Read market and account data from Hyperliquid",
-		Long: `Query the Hyperliquid Info API for market data, order books, trades,
+	return newHelpCommandGroup(
+		"info",
+		"Read market and account data from Hyperliquid",
+		`Query the Hyperliquid Info API for market data, order books, trades,
 candles, funding rates, account state, and open orders. All info commands
 are read-only and require no wallet configuration.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return cmd.Help()
-		},
-	}
-
-	cmd.AddCommand(
 		newInfoLookupCmd(),
 		newInfoMidsCmd(),
 		newInfoMetaCmd(),
@@ -31,6 +25,4 @@ are read-only and require no wallet configuration.`,
 		newInfoFundingCmd(),
 		newInfoPerpDexsCmd(),
 	)
-
-	return cmd
 }
