+++ cmd/order_schedule_cancel.go
@@ -18,11 +18,9 @@ func newOrderScheduleCancelCmd() *cobra.Command {
 or clear an existing schedule. Exactly one of --timeout or --clear must be provided.`,
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			cfg := config.FromContext(cmd.Context())
-			timeoutStr, _ := cmd.Flags().GetString("timeout") //nolint:errcheck // known flag
-			clear, _ := cmd.Flags().GetBool("clear")          //nolint:errcheck // known flag
+			clear, _ := cmd.Flags().GetBool("clear") //nolint:errcheck // known flag
 
 			hasTimeout := cmd.Flags().Changed("timeout")
-
 			if hasTimeout && clear {
 				return output.NewCLIError(output.ErrValidation, "--timeout and --clear are mutually exclusive")
 			}
@@ -30,36 +28,18 @@ or clear an existing schedule. Exactly one of --timeout or --clear must be provi
 				return output.NewCLIError(output.ErrValidation, "one of --timeout or --clear is required")
 			}
 
-			var cancelTime *int64
-			if hasTimeout {
-				d, err := time.ParseDuration(timeoutStr)
-				if err != nil {
-					return output.NewCLIError(output.ErrValidation, "invalid timeout duration").
-						WithDetails("value", timeoutStr).
-						WithDetails("hint", "use Go duration format (e.g. 5m, 1h, 30s)")
-				}
-				if d <= 0 {
-					return output.NewCLIError(output.ErrValidation, "timeout must be positive").
-						WithDetails("value", timeoutStr)
-				}
-				if d < 5*time.Second {
-					return output.NewCLIError(output.ErrValidation, "timeout must be at least 5 seconds").
-						WithDetails("value", timeoutStr)
-				}
-				ms := time.Now().Add(d).UnixMilli()
-				cancelTime = &ms
+			cancelTime, err := parseScheduleCancelTime(cmd, hasTimeout)
+			if err != nil {
+				return err
 			}
-			// When --clear, cancelTime stays nil (which tells API to remove the schedule).
 
 			exec, err := buildExecutor(cfg)
 			if err != nil {
 				return err
 			}
 
-			result, err := exec.ScheduleCancel(cmd.Context(), exchange.ScheduleCancelInput{
-				Time:   cancelTime,
-				DryRun: cfg.DryRun,
-			})
+			input := exchange.ScheduleCancelInput{Time: cancelTime, DryRun: cfg.DryRun}
+			result, err := exec.ScheduleCancel(cmd.Context(), input)
 			if err != nil {
 				return err
 			}
@@ -72,4 +52,29 @@ or clear an existing schedule. Exactly one of --timeout or --clear must be provi
 	cmd.Flags().Bool("clear", false, "clear existing schedule")
 
 	return cmd
+}
+
+func parseScheduleCancelTime(cmd *cobra.Command, hasTimeout bool) (*int64, error) {
+	if !hasTimeout {
+		return nil, nil
+	}
+
+	timeoutText, _ := cmd.Flags().GetString("timeout") //nolint:errcheck // known flag
+	timeout, err := time.ParseDuration(timeoutText)
+	if err != nil {
+		return nil, output.NewCLIError(output.ErrValidation, "invalid timeout duration").
+			WithDetails("value", timeoutText).
+			WithDetails("hint", "use Go duration format (e.g. 5m, 1h, 30s)")
+	}
+	if timeout <= 0 {
+		return nil, output.NewCLIError(output.ErrValidation, "timeout must be positive").
+			WithDetails("value", timeoutText)
+	}
+	if timeout < 5*time.Second {
+		return nil, output.NewCLIError(output.ErrValidation, "timeout must be at least 5 seconds").
+			WithDetails("value", timeoutText)
+	}
+
+	cancelTime := time.Now().Add(timeout).UnixMilli()
+	return &cancelTime, nil
 }
