+++ pkg/signer/eip712.go
@@ -30,23 +30,13 @@ var eip712DomainType = []apitypes.Type{
 // l1Domain returns the EIP-712 domain for L1 phantom agent signing.
 // name="Exchange", version="1", chainId=1337, verifyingContract=0x000...000.
 func l1Domain() apitypes.TypedDataDomain {
-	return apitypes.TypedDataDomain{
-		Name:              "Exchange",
-		Version:           "1",
-		ChainId:           math.NewHexOrDecimal256(chainIDL1),
-		VerifyingContract: zeroAddress,
-	}
+	return apitypes.TypedDataDomain{Name: "Exchange", Version: "1", ChainId: math.NewHexOrDecimal256(chainIDL1), VerifyingContract: zeroAddress}
 }
 
 // userActionDomain returns the EIP-712 domain for user-signed actions.
 // name="HyperliquidSignTransaction", version="1", chainId=421614, verifyingContract=0x000...000.
 func userActionDomain() apitypes.TypedDataDomain {
-	return apitypes.TypedDataDomain{
-		Name:              "HyperliquidSignTransaction",
-		Version:           "1",
-		ChainId:           math.NewHexOrDecimal256(chainIDUserSigned),
-		VerifyingContract: zeroAddress,
-	}
+	return apitypes.TypedDataDomain{Name: "HyperliquidSignTransaction", Version: "1", ChainId: math.NewHexOrDecimal256(chainIDUserSigned), VerifyingContract: zeroAddress}
 }
 
 // hyperliquidChain returns the chain name string for EIP-712 message content.
@@ -66,13 +56,5 @@ func userActionTypedData(typeName string, typeFields []apitypes.Type, message ma
 	maps.Copy(msg, message)
 	msg["hyperliquidChain"] = hyperliquidChain(isMainnet)
 
-	return apitypes.TypedData{
-		Types: apitypes.Types{
-			"EIP712Domain": eip712DomainType,
-			typeName:       typeFields,
-		},
-		PrimaryType: typeName,
-		Domain:      userActionDomain(),
-		Message:     msg,
-	}
+	return apitypes.TypedData{Types: apitypes.Types{"EIP712Domain": eip712DomainType, typeName: typeFields}, PrimaryType: typeName, Domain: userActionDomain(), Message: msg}
 }
